+++ dlfnext-20250310/core/Models/Share.cs
@@ -21,6 +21,10 @@ namespace AlibabaCloud.SDK.DlfNext20250310.Models
         [Validation(Required=false)]
         public string CreatedBy { get; set; }
 
+        [NameInMap("enableWrite")]
+        [Validation(Required=false)]
+        public bool? EnableWrite { get; set; }
+
         [NameInMap("owner")]
         [Validation(Required=false)]
         public string Owner { get; set; }